Move `FeatureFlag` to the toplevel
This fixes the compile errors for some reason ?!?!?
